requirements.txt
mistralai==0.0.12
numpy==1.26.4
openai==1.12.0
openpyxl==3.1.2
pandas==2.2.0
pdfservices-sdk==2.3.0
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0
regex==2023.12.25
requests==2.27.1
scipy==1.12.0
tiktoken==0.6.0
tokenizers==0.15.1
torch==2.2.0
tqdm==4.66.2
transformers==4.37.2
virtualenv==20.25.0